Introduce action to bump and tag versions in monorepos
This commit introduces a GitHub action that can be used to
automatically bump and tag versions of components that live in Git
monorepos.

For example, if we have a repository that contains two components,
say "component1" and "component2", who live in the subdirectories
"components/component1" and "components/component2", respectively,
the action can be used to tag commits that modify one of the
components with a new version number.

The format for tags is `COMPONENT-vVERSION`, where `COMPONENT` is
the name of the component in kebab-case, and `VERSION` is the
version number in semantic versioning format.

The action bumps the version number by traversing backwards across
all commits that modify the relevant component until it reaches
the previously tagged commit. All commit messages after the
previously tagged commit are checked for the hashtags `#major`,
`#minor` and `#patch` in order to decide which version component
to bump. The highest version component wins (so, if there are
multiple commits modifying the component, the commit with the
highest hashtag wins, major>minor>patch). If no hashtags
exist, the minor component is bumped. If no previous tag is found,
a starting version is used, which can either be provided from the
calling workflow, or defaults to 0.0.0.

The action includes unit tests and a README file with usage and
development information.